requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
python-dotenv>=1.0.1
pytest>=8.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
jinja2>=3.1.3
python-multipart>=0.0.9
apscheduler>=3.10.4
